The first interation of the game
This is the first iteration of the game so far. The player has a chance of coming in contact with 4 different enemies. Honestly the bear is over powered but like in real life bears are the most apex predator. Next will add exp and drops like items which will involve an addition of and items class.